moved operator functions to the operator functions file renamed generic file to items file
